Solving problems with the type (day, night) in getWeekType and performance detection cycles when switching between mods. Issue 1: The getWeekType(weekNumber) function did not work correctly in short-long-mode.
Solution 1: We tried changing the logic of the function using the sequence list length instead of cycleLength, but it didn't help.

Issue 2: The getWeekType(weekNumber) function could not determine the week type based on the day of the week.

Solution 2: We tried using a different logic, using weekNumber to determine the day of the week and then determining the type of week based on that day.

Issue 3: The getWeekType(weekNumber) function could not determine the week type based on the cycle.

Solution 3: We found a solution by using cycleLength to determine the length of the cycle and then using weekCycleIndex to determine the type of week.

Result: The getWeekType(weekNumber) function now works correctly and can determine the week type based on the cycle.